refactor: minor refactoring to package structure and remove unused functions
