+++ plugin/compiler-plugin/src/main/kotlin/it/unibo/collektive/utils/call/CallAlignment.kt
@@ -2,8 +2,6 @@ package it.unibo.collektive.utils.call
 
 import it.unibo.collektive.utils.common.getAlignmentToken
 import it.unibo.collektive.utils.common.getLambdaType
-import it.unibo.collektive.utils.common.putTypeArgument
-import it.unibo.collektive.utils.common.putValueArgument
 import it.unibo.collektive.utils.stack.StackFunctionCall
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.backend.common.lower.createIrBuilder
@@ -41,17 +39,17 @@ fun IrSingleStatementBuilder.buildAlignedOnCall(
     data: Map<String, Int>,
 ): IrFunctionAccessExpression {
     return irCall(alignedOnFunction).apply {
-        // Set generics type
-        putTypeArgument(expression.type)
+        // Set the return type
+        type = expression.type
+        // Set generics type of the `alignOn` function
+        putTypeArgument(0, type)
         // Set aggregate context
         putArgument(alignedOnFunction.dispatchReceiverParameter!!, aggregateContextReference)
         // Set the argument that is going to be push in the stack
         val functionName = expression.getAlignmentToken()
         val count = data[functionName]!! // Here the key should be present!
         val alignmentToken = "$stack$functionName.$count"
-        putValueArgument(
-            irString(alignmentToken),
-        )
+        putValueArgument(0, irString(alignmentToken))
         // Create the lambda that is going to call expression
         val lambda = buildLambdaArgument(pluginContext, aggregateLambdaBody, expression)
         putValueArgument(1, lambda)
